Rename wrapper to normalizer
