+++ src/Sylius/Bundle/PaymentBundle/Resources/config/services/controllers.xml
@@ -21,7 +21,7 @@
 
         <service id="sylius.controller.payment_request_notify" class="Sylius\Bundle\PaymentBundle\Action\PaymentRequestNotifyAction">
             <argument type="service" id="sylius.repository.payment_request" />
-            <argument type="service" id="sylius.wrapper.symfony_request" />
+            <argument type="service" id="sylius.normalizer.symfony_request" />
             <argument type="service" id="sylius_abstraction.state_machine" />
             <argument type="service" id="sylius.announcer.payment_request" />
             <argument type="service" id="sylius.manager.payment_request" />
@@ -30,7 +30,7 @@
         <service id="sylius.controller.payment_method_notify" class="Sylius\Bundle\PaymentBundle\Action\PaymentMethodNotifyAction">
             <argument type="service" id="sylius.repository.payment_method" />
             <argument type="service" id="sylius.factory.payment_request" />
-            <argument type="service" id="sylius.wrapper.symfony_request" />
+            <argument type="service" id="sylius.normalizer.symfony_request" />
             <argument type="service" id="sylius.repository.payment_request" />
             <argument type="service" id="sylius.announcer.payment_request" />
             <argument type="service" id="sylius.provider.payment_notify" />
