+++ src/Sylius/Bundle/PaymentBundle/Resources/config/services/normalizer.xml
@@ -16,6 +16,6 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
 >
     <services>
-        <service id="sylius.wrapper.symfony_request" class="Sylius\Bundle\PaymentBundle\Wrapper\SymfonyRequestWrapper" />
+        <service id="sylius.normalizer.symfony_request" class="Sylius\Bundle\PaymentBundle\Normalizer\SymfonyRequestNormalizer" />
     </services>
 </container>
